Rearrange images and refactor paths
