+++ blog-site/src/components/footer.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Link } from 'gatsby'
 import { rhythm, scale } from '../utils/typography'
-import logo from './DOI-2x.png'
-import twitter from './twitter.png'
-import facebook from './facebook.png'
-import github from './github.png'
+import logo from '../img/DOI-2x.png'
+import twitter from '../img/twitter.png'
+import facebook from '../img/facebook.png'
+import github from '../img/github.png'
 
 const Footer = () => (
   <footer
