+++ blog-site/src/components/layout.js
@@ -4,7 +4,7 @@ import { Link } from 'gatsby'
 import { rhythm, scale } from '../utils/typography'
 import './layout.css'
 import Footer from './footer.js'
-import mastImage from './NRRD_blog_mast.png'
+import mastImage from '../img/NRRD_blog_mast.png'
 import Banner from './govbanner.js'
 class Template extends React.Component {
