+++ blog-site/src/data/author.yaml
@@ -1,23 +1,23 @@
 - id: Ryan Johnson
   bio: content strategist at the Office of Natural Resources Revenue
-  pic: /img/ryan-pic.jpg
+  pic: ../img/ryan-pic.jpg
 
 - id: Jennifer Malcolm
   bio: product manager at the Office of Natural Resources Revenue
-  pic: /img/jennifer-pic.jpg
+  pic: ../img/jennifer-pic.jpg
 
 - id: Chris Mentasti
   bio: program analyst at the Office of Natural Resources Revenue
-  pic: /img/chris-pic.jpg
+  pic: ../img/chris-pic.jpg
 
 - id: Shannon McHarg
   bio: user experience designer at the Office of Natural Resources Revenue
-  pic: /img/shannon-pic.jpg
+  pic: ../img/shannon-pic.jpg
 
 - id: Jeff Keene
   bio: front end developer at the Office of Natural Resources Revenue
-  pic: /img/jeff-pic.jpg
+  pic: ../img/jeff-pic.jpg
 
 - id: Lindsay Goldstein
   bio: program analyst at the Office of Natural Resources Revenue
-  pic: /img/lindsay-pic.png
+  pic: ../img/lindsay-pic.png
